@@ -1,7 +1,7 @@
 ﻿@using System.Collections.Generic
 @using System.Linq
 @using Microsoft.AspNetCore.Http
-@using Microsoft.AspNetCore.Http.Authentication
+
 @model LoginViewModel
 @inject SignInManager<ApplicationUser> SignInManager
 
